file renaming and import updates
