+++ lib/utilities/apiWrapper.ts
@@ -1,5 +1,5 @@
 
-import fetch from './retry'
+import fetch from './fetch'
 
 import { BUTTER_BASE_API_URL, BUTTER_BASE_HEADERS } from '../config'
 import type { GlobalButterConfig } from '../typescript/GlobalButterConfig'
